feat: add sonic utils to parse JSON
