+++ pkg/render/json.go
@@ -1,0 +1,9 @@
+package render
+
+//type JSONWriter interface {
+//	Render(data []byte) (string, error)
+//}
+//
+//type JSONBuilder struct {
+//	builder strings.Builder
+//}
